% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveTempGrid.R
\name{effectiveTempGrid}
\alias{effectiveTempGrid}
\title{Effective temperature}
\usage{
effectiveTempGrid(tasmin = NULL, tasmax = NULL, k = 0.69)
}
\arguments{
\item{tasmin}{Grid of minimum monthly temperature (degC)}

\item{tasmax}{Grid of maximum monthly temperature (degC)}

\item{k}{Calibration coefficient. Default to 0.69, as proposed by Pereira and Pruitt (2004)}
}
\description{
Calculation of the \dQuote{effective temperature} for improving the Thornthwaite's ET0 estimation method
}
\details{
The function is internally used by \code{\link{petGrid}} when the Thornthwaite's method with calibration coefficient k is chosen.
}
\references{
\itemize{
\item de Camargo, A.P., Marin, F.R., Sentelhas, P. C., Picini, A.G., 1999. Adjust of the Thornthwaite’s method to estimate the potential evapotranspiration for arid and superhumid climates, based on daily temperature amplitude. Revista Brasileira de Agrometeorologia 7.
\item Pereira, A.R., Pruitt, W.O., 2004. Adaptation of the Thornthwaite scheme for estimating daily reference evapotranspiration. Agricultural Water Management 66, 251–257. https://doi.org/10.1016/j.agwat.2003.11.003
}
}
\seealso{
\code{\link{petGrid}}
}
\author{
J. Bedia
}
