% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petGrid.R
\name{petGrid}
\alias{petGrid}
\title{Calculation of Potential Evapotranspiration}
\usage{
petGrid(tasmin = NULL, tasmax = NULL, tas = NULL, pr = NULL,
  method = c("thornthwaite", "hargreaves"))
}
\arguments{
\item{tasmin}{Grid of minimum monthly temperature (degC)}

\item{tasmax}{Grid of maximum monthly temperature (degC)}

\item{tas}{Grid of mean monthly temperature (degC)}

\item{pr}{Grid of total monthly precipitation amount (mm/month)}

\item{method}{Potential evapotranspiration method. Currently \code{"thornthwaite"} and 
\code{"hargreaves"} methods are available. See details.}
}
\description{
Implementation of several PET methods
}
\details{
This function is a wrapper of the functions with the same name as given in \code{method}
from the \pkg{SPEI} package (Begueria and Vicente-Serrano). Monthly input data are thus required.
}
\author{
J Bedia
}
