% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{tasmax.cru.iberia}
\alias{tasmax.cru.iberia}
\title{Mean maximum monthly temperature data in the Iberian Peninsula 1981-2010}
\format{
A \pkg{climate4R} grid
}
\source{
Climate Research Unit, University of East Anglia (CRU TS v4.00), \url{https://crudata.uea.ac.uk/cru/data/hrg/}.
The R dataset has been created with package \pkg{loadeR}, of the \strong{climate4R bundle} (Cofino et al 2017)
}
\description{
Gridded observations of monthly mean maximum surface temperature for the Iberian Peninsula (1981-2010, in degC)
}
\examples{
data("tasmax.cru.iberia")
require(visualizeR)
spatialPlot(climatology(tasmax.cru.iberia), backdrop.theme = "countries", rev.colors = TRUE)
}
\references{
\itemize{
\item Harris, I., Jones, P.D., Osborn, T.J., Lister, D.H., 2014. Updated high-resolution grids of monthly climatic observations - the CRU TS3.10 Dataset: UPDATED HIGH-RESOLUTION GRIDS OF MONTHLY CLIMATIC OBSERVATIONS. International Journal of Climatology 34, 623-642. doi:10.1002/joc.3711
\item Cofino, A.S., Bedia, J., Iturbide, M., Vega, M., Herrera, S., Fernandez, J., Frias, M.D., Manzanas, R., Gutierrez, J.M., 2017. The ECOMS User Data Gateway: Towards seasonal forecast data provision and reseacrh reproducibility in the era of climate services. Climate Services. doi:10.1002/joc.3711
}
}
