% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petGrid.R
\name{petGrid}
\alias{petGrid}
\title{Calculation of Potential Evapotranspiration}
\usage{
petGrid(tasmin = NULL, tasmax = NULL, tas = NULL, pr = NULL,
  method = c("thornthwaite", "hargreaves"), ...)
}
\arguments{
\item{tasmin}{Grid of minimum monthly temperature (degC)}

\item{tasmax}{Grid of maximum monthly temperature (degC)}

\item{tas}{Grid of mean monthly temperature (degC)}

\item{pr}{Grid of total monthly precipitation amount (mm/month)}

\item{method}{Potential evapotranspiration method. Currently \code{"thornthwaite"} and 
\code{"hargreaves"} methods are available. See details.}

\item{...}{Further arguments passed to the PET internals}
}
\description{
Implementation of several PET methods for the climate4R bundle
}
\details{
This function is a wrapper of the functions with the same name as given in \code{method}
from the \pkg{SPEI} package (Begueria and Vicente-Serrano). Monthly input data are thus required. 
The latitude of the sites/grid points is internally used.
In case of multimember grids (e.g. seasonal forecast data), the PET is calculated for each member sepparately.
}
\examples{
# Thorthwaite requires monthly mean temperature data as input:
data("tas.cru.iberia")
thpet <- petGrid(tas = tas.cru.iberia, method = "thornthwaite")
require(transformeR)
require(magrittr)
# This is the climatology (by seasons)
seasons <- list("DJF" = c(12,1,2), "MAM" = 3:5, "JJA" = 6:8, "SON" = 9:11)
season.list <- lapply(1:length(seasons), function(x) {
    subsetGrid(thpet, season = seasons[[x]]) \%>\% 
               aggregateGrid(aggr.y = list(FUN = "sum")) \%>\% climatology()
})
mg <- makeMultiGrid(season.list, skip.temporal.check = TRUE)
plotClimatology(mg, names.attr = names(seasons), at = seq(0, 525, 10),
                backdrop.theme = "coastline",
                main = "Mean Potential Evapotranspiration Thornthwaite (1981-2010, mm/season)")
# A typical operation is computing trends
# Here we are interested in the JJA PET trend in Iberia:
# We consider the annually aggregated PET series (n = 30 years)
jja.pet <- subsetGrid(thpet, season = 6:8) \%>\% aggregateGrid(aggr.y = list(FUN = "sum"))
trend.thpet <- climatology(jja.pet, clim.fun = list(FUN = "trend.1D",
                                                    dates = getRefDates(jja.pet),
                                                    method = "kendall"))
plotClimatology(trend.thpet, backdrop.theme = "countries",
                main = "Mann-Kendall Tau PET trend (JJA, 1981-2010)")
pval.estimate <- climatology(jja.pet, clim.fun = list(FUN = "trend.1D",
                                                      dates = getRefDates(jja.pet),
                                                      method = "kendall",
                                                      return.pvalue = TRUE))
sig.points <- map.stippling(clim = pval.estimate, threshold = 0.05, condition = "LT", 
                            pch = 19, cex = .5, col = "purple")
plotClimatology(trend.thpet, backdrop.theme = "countries",
                main = "Mann-Kendall Tau PET trend (JJA, 1981-2010)",
                sp.layout = list(sig.points))
# See further examples in 'help(trend.1D)'
}
\author{
J Bedia
}
