% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speiGrid.R
\name{speiGrid}
\alias{speiGrid}
\title{Computation of Standardized Precipitation(-Evapotranspiration) Index Grids}
\usage{
speiGrid(pr.grid, et0.grid = NULL, scale = 3, ...)
}
\arguments{
\item{pr.grid}{Precipitation grid (monthly accumulated values, in mm)}

\item{et0.grid}{Potential evapotranspiration grid (same units as \code{pr.grid})}

\item{scale}{Integer. Time scale at which the SPEI/SPI are computed. Default to 3 (months)}

\item{...}{Further arguments passed to \code{\link[SPEI]{spei}}}
}
\value{
A \strong{climate4R} grid with SPEI/SPI data
}
\description{
Returns SPI/SPEI grid using precipitation (and PET for SPEI) input grids
}
\details{
The function is a wrapper of function \code{\link[SPEI]{spei}} from package \pkg{SPEI} adapted
to \strong{climate4R} input grids
}
\examples{
# By default, et0.grid is null, and SPI is computed from precipitation: 
data("pr.cru.iberia")
spi3 <- speiGrid(pr.grid = pr.cru.iberia, scale = 3)
## If PET is used, then SPEI is calculated
data("tas.cru.iberia")
et0.grid <- petGrid(tas = tas.cru.iberia, method = "thornthwaite")
spei3 <- speiGrid(pr.cru.iberia, et0.grid = et0.grid, scale = 3)
}
\seealso{
\code{\link{petGrid}}, for PET calculation
}
